[CL-1110] Refactor for new backend data structures (WIP)
